Admin can now block and unblock customers

+++ app/Http/Controllers/AnnouncementController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Carbon;
@@ -10,36 +11,49 @@ use App\Models\Announcement;
 
 class AnnouncementController extends Controller
 {
-    public function viewAnnouncements(){
-        $announcements = Announcement::latest()->paginate('5');
+    public function viewAnnouncements(Request $request){
+        $announcements = Announcement::latest()->paginate(5);
         if(Auth::check() && Auth::user()->account_type == 'admin'){
             return view('admin/admin-home');
         }else if(Auth::check() && Auth::user()->account_type == 'customer'){
-            return view('home',compact('announcements'));
+            if(Auth::user()->account_status == 'active'){
+                return view('home',compact('announcements'));
+            }else{
+                Auth::logout();
+                session()->flash('success','Your account has been blocked by the Administrator.');
+                return view('auth/login');
+            }
         }else{
+            Auth::logout();
             return view('guest-home',compact('announcements'));
         }
     }
 
     public function adminViewAnnouncements(){ //exclusively for admin side..
-        $announcements = Announcement::latest()->paginate('5');
+        $announcements = Announcement::latest()->paginate(5);
         if(Auth::check() && Auth::user()->account_type == 'admin'){
             return view('admin/admin-manageAnnouncements',compact('announcements'));
         }else{
+            Auth::logout();
             return view('auth/login');
         }
     }
 
     public function createAnnouncement(Request $request){
-        $validated = $request->validate([
-            'announcementContent' => 'required|string',
-        ]);
+        if(Auth::check() && Auth::user()->account_type == 'admin'){
+            $validated = $request->validate([
+                'announcementContent' => 'required|string',
+            ]);
 
-        Announcement::create([
-            'announcement_content' => $request->announcementContent,
-            'created_at' => Carbon::now()
-        ]);
-        return Redirect()->route('admin-manageAnnouncements')->with(['success'=>'Announcement Created Successfully!']);
+            Announcement::create([
+                'announcement_content' => $request->announcementContent,
+                'created_at' => Carbon::now()
+            ]);
+            return Redirect()->route('admin-manageAnnouncements')->with(['success'=>'Announcement Created Successfully!']);
+        }else{
+            Auth::logout();
+            return view('auth/login');
+        }
     }
 
     //the function below redirects the admin to the 'edit announcement form' for a specific announcement
@@ -48,22 +62,33 @@ class AnnouncementController extends Controller
             $announcement = Announcement::find($request->id);
             return view('admin/admin-editAnnouncement',compact('announcement'));
         }else{
+            Auth::logout();
             return view('auth/login');
         }
     }
 
     //the function below will be the one to update the announcement from the mysql database after form submission..
     public function updateAnnouncement(Request $request){
-        $update = Announcement::find($request->announcement_id)->update([
-            'announcement_content'=> $request->announcementContent
-        ]);
-        return Redirect()->route('admin-manageAnnouncements')->with(['success'=>'Announcement Updated Successfully!']);
+        if(Auth::check() && Auth::user()->account_type == 'admin'){
+            $update = Announcement::find($request->announcement_id)->update([
+                'announcement_content'=> $request->announcementContent
+            ]);
+            return Redirect()->route('admin-manageAnnouncements')->with(['success'=>'Announcement Updated Successfully!']);
+        }else{
+            Auth::logout();
+            return view('auth/login');
+        }
     }
 
     public function deleteAnnouncement(Request $request){
-        $announcementData = Announcement::find($request->announcement_id);
-        $announcementData->delete();
-        return Redirect()->route('admin-manageAnnouncements')->with(['success'=>'Announcement Removed Successfully!']);
+        if(Auth::check() && Auth::user()->account_type == 'admin'){
+            $announcementData = Announcement::find($request->announcement_id);
+            $announcementData->delete();
+            return Redirect()->route('admin-manageAnnouncements')->with(['success'=>'Announcement Removed Successfully!']);
+        }else{
+            Auth::logout();
+            return view('auth/login');
+        }
     }
     
 }
